refactor: develop branch 업데이트 사항 적용
- stillcut -> snapshot 이름 변경

Refs: #70

+++ src/main/java/com/movte/slate/community/api/request/ViewOtherUserFeedsApiRequest.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 public class ViewOtherUserFeedsApiRequest {
     private long otherUserId;
     private long lastFeedId;
-    private long pageSize;
+    private int pageSize;
 
     public ViewOtherUserFeedsServiceRequest toServiceRequest() {
         return new ViewOtherUserFeedsServiceRequest(otherUserId, lastFeedId, pageSize);
